Add `Spok::Calendars` API
This module will be responsible for dealing with our collection of
builtin calendars and provide a way for users to add their own calendars
inside their app, without having to add them to the gem and require new
releases to RubyGems.

The existing calendars are still here, but in the future we could deprecate
them and remove them from the codebase, or tweak to code to avoid loading
all 24 files into memory.